reedme
we do: 
you can read a peace of diolog with one big clouthing catalogue

we sell products on our sites as a new brand. I will find an example of the site below.
 we will have a higher conversion than about Yoins because in Russia is not popular products with shipping.
We on the site we will sell purchases in installments (by partners with Admitad) 
we better know the mentality of Russian girls and males. 
We also give an affectionate support service that will allow the user to shop by phone.
Starting to fullfillment  Yoins.
Our mission is the maximum balance and maximum respect of our customers, by doing this we need quality control of the delivered goods + control at each stage. All we need is onlinne integretion  and  perfect suuport from Yoins.
We must also understand that we do not succeed in working according to standard conditions (as a web master), We-Admitad-Yoins, where we are a fullfilm, Admitad is a technical consultant and "Bank" and guarantor - Yoins is a manufacturer. And this will become pogromous when we go to all Russian Wildberries, Azon and so on everything on our market they imply there six days in my opinion up to six days delivery so it will be necessary to have goods in Russia and having on hand three third party contract We-Admitad-Yoins - it will be completely safe and everyone will earn.